requirements.txt
# Core Dependencies
python-dateutil>=2.8.2
pytz>=2023.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Clients
requests>=2.31.0
openai>=1.0.0

# News Fetching
feedparser>=6.0.10

# Dashboard & UI
streamlit>=1.28.0
plotly>=5.17.0

# Configuration
python-dotenv>=1.0.0

# Utilities
loguru>=0.7.0
